Restored Auto Generate icon and saving desktop icons.
Can now auto generate icons once again and all issues which should stop a save of a desktopIcon should be fixed.

Known issue: Icon not re-rendering after changing, and not displaying preview before saving. For example it needs a restart to update its icon. These just need the functionality re-adjusted (not a bug just not implemented).

Also not showing edit mode(This one is not even attempted at implementing yet).