grid_menu improved placement
Now attempts to place the Menu on the side with more room (keep the menu centered). And fixed it such that if part of the menu would appear below the window (editing a desktopIcon near the bottom) it now adjusts the window's Y location to keep it within the window.

Also fixed a bug where updating the Icon Size in settings was not updating global ICON_SIZE variable (Thus would cause problems like Menu appearing based on the old ICON_SIZE not the new adjusted one).