Fix changing max rows/cols removing background video/breaking image
Fixed an unintended bug of removing the video background upon changing max rows/columns.

This also seemed to fix a bug where changing the max rows/columns in a non-maximized window, would then upon making the window larger, instead of scaling the image to fit the screen would instead do a grid texture of smaller versions of the background in the newly added area.