Restored dropping files on icons -> Edit menu
Restored the ability from the old grid to drop a file on an icon to open the Edit Icon menu with it set as the Icon path (if file is an icon) or the Executable path (all other filetypes).

Also directly placed the MEDIA_PLAYER pausing and unpausing inside the show_grid_menu() function instead of calling before and after in the now 3 instances where it calls for showing the grid_menu.