Fix background video not playing issue.
Fixed an issue which would lead to the video background not playing upon changing to video background.

Was caused by a function meant to not reload the video (and restart the video from the start) every time a redraw was requested. Thus it was not loading the video into the media player. It now checks if a source exists for the media player and adds the source if it does not exist.

Also removed a function which simply caused the program to reload global variables twice.