More control for updating from older versions. (Skipping versions)
A problem was created with V0.2.000 where a quick trick was done to flip the setting since I re-organized the "On closing the window" setting. The problem is that it worked assuming everyone would update from V0.1.002 to V0.2.000. However you can't just do the same trick and flip the bit every update following that.

So instead the process is to now keep track of a setting called "updated from" which tells us the version the user updated from. We can then read from that when showing patch notes (using the code where patch notes are only shown upon updating), to call a function which depending on "updated_from" can change the setting etc.

For reference any version <= V0.2.000 upon updating will reset the "On closing the program" setting to "Terminate the program" upon updating. Updating from version V0.2.001 or beyond would not reset to "Terminate the program" but would keep the users settings the same.