New setting: show "+" icon on default icon hover
New setting added to the settings menu.
"Show + icon on default hover"
This controls whether or not hovering over an empty icon shows the add.png "+" looking icon.

This add.png icon mostly exists to show a new user that they could double click/right click on these to add icons.
But is pretty annoying for someone who knows the gist of how to use the program, thus a setting has been added to remove it.