Pausing video playback on settings/contextmenu
Video playback can be quite taxing. Eventually I think it needs to be rewritten either with threading or something to be more efficient as it can lag the entire program while the video is playing (this doesn't affect it minimized etc. just while playing).

For now at least it will now pause this playback while working with most menus/occasions where the user might notice the lag.

Pausing the video player now happens on right click (showcontextmenu) and in the settings menu.