proof of concept: updating icon name font size.
For updating font size in general a lot more work needs to be done. For instance, some work behind the scenes with multi line commenting scaling based on the size of the font to fit into the allocated space.

For now added a test to scroll up/down which was used early on to debug between 2 different icon sizes (which may have been forgotten...). For now all it does is change the font size for the [0][0] icon (1st icon in the very top left corner) between 18 and 10 (the default size).

Like stated in title this is just a proof of concept. The idea eventually is to have a set font and size for all new icons (with the possibility to override all in the settings menu) and have setting icons name size/font individually as well.