Breaking everything to move forward.
Recap as to what I have been having the problem with:
I have found a bug where based on the size (pixel dimensions) of the video, you will reach a point where for seemingly no logical reason the desktopIcons will start an infinite loop calling paint() on themselves as fast as possible. While it was never a memory leak or a cpu eating machine (~5-6% cpu usage max), it would lag the hell out of the video player and lag the menus. It is not even something simple like just return upon calling paint() because it is actually infinite and it lags not due to re-painting but literally because it is calling it as fast as it can. The only solution to not getting this infinite loop was to have a smaller video(window) size or keep the rows/columns under a total of ~230 total desktop icons(>~270 with a smaller screen etc.) INCLUDING the blank icons.

I have spent ~5 hours trying multiple different video libraries. (ffmpeg, cv2, av, QVideoWidget etc.) Attempted multiple different layouts/formatting trying to get the icons to render on top of the video to no avail.

I believe I have finally found a good way forward. On the plus side the solution at least seems like it will be more optimal in all situations.

The way forward seems to be to just ditch the whole create row * column number of desktopIcons and instead only create the desktopIcons for icons which exist(non-default). I originally did it just for the easiness to allow drops on the individual icons and for showing the "+" icon on hover. I should be able to re-implement the drops in just the QGraphicsView eventually, unsure about the "+" icons.

Known issues (things I broke):
Icon resizing in settings menu.
self.desktop_icons still is [row][col] size (but even not creating the desktopIcons fixed the problems with infinite paint() calls).
self.desktop_icons should instead only be as big as icons that exist (probably not even 2d array but just a normal array).
Cheap tricks for icon_visiblities which just check self.desktop_icons[x][y] is not none.
Almost all dropping/moving icons is broken (besides 2 working icons) or dropping a file on an existing icon.

I'm 100% sure other stuff is broken :), but I will try to find it before next official release