desktop_icons is now a dictionary
Changed the storage of self.desktop_icons to a dictionary instead of a 2d array. Which only stores existing desktop_icons. This should be much more efficient as it only stores the icons which exist.

Regardless of whether I kept the 2d array or not I would have to implement a way for right click/hover etc to create "fake" desktopIcons either way so I might as well use a more efficient format for accessing and storing them.

As the entire formatting has changed and it should be a lot more efficent to loop through all set icons I have removed the slightly more efficient version of update_icon_visibility and replaced it with what was originally update_fresh_icon_visibility (the slightly less efficient version for running at startup). Since as a whole the efficiency should be increased by using a dictionary instead of a 2d array it should still be a performance increase.

Still have multiple things to fix like dropping, swapping icons, creating icons (from fresh non-existing icons). etc.