Mouse doubleclick passthrough and control
One of the steps to restoring ability to double click and right click on blank/empty icons is moving the individual mouseDoubleClickEvent to the QGraphicsView, to get a row/col. Then we can check if that icon exist and call a (now renamed) double_click event to mimic the original functionality of the desktop icons mouseDoubleClickEvent, while allowing the use of a double click on blank icons to eventually call show_grid_menu() for their position.