Fixed bug on creating new icon.
Upon saving from grid_menu now adds the icon to self.desktop_icons and fixes a bug which tried to reload it when it didn't exist in self.desktop_icons.

Also fixed a small bug when saving a desktop icon which compared self.exec_path_le instead of self.exec_path_le.text() causing it to display a warning for an invalid executable path even though the path is empty.