Now draws red border around blank icons in edit mode
Added RedBorderItem a temporary item which when edit_mode_icon is called if the icon is not already in self.desktop_icons (Icon is a default icon) it will add a RedBorderItem at the location which mimics the red border drawn from a DesktopIcon when in edit mode.

The only difference is that it does not account for text area, but just draws around where the icon itself would go.

Item gets removed upon calling normal_mode_icon() which is called upon saving/closing grid_menu.