ContextMenu moved to QGraphicsView
To be able to right click and see a context menu on non-existing desktop icons QGraphicsView must hold the contextMenuEvent.

For right clicking on existing icons it now calls the icon's context_menu (renamed from contextMenuEvent).

So the functionality for now should be the same as the last commit version, but it paves the way to create a context menu for NON desktop icon spaces.