dropEvent now goes through QGraphicsView.
In order to implement the dropEvent for non-existing icons it needs to be used by the QGraphicsView.

Dropping an icon on an existing icon retains functionality (dropEvent works as a pass-thru calling drop_event for DesktopIcon).

Swapping icons still works. Swapping an icon and a non-icon is still broken. And dropping a file on a non-existing icon does not yet do anything.