Fixed bug where multiple RedBorderIcons could be created
edit_mode_icon now removes an existing self.red_border_icon if it exists when it is called.

I had found a bug which was causing edit_mode_icon to be called twice upon dropping a file on a blank icon. This lead to the creation of a second red_border_icon which no longer had a reference and was not removed from the scene.

Upon going into edit mode now, a blank icon will delete an existing red_border_icon should it exist before creating a new red_border_icon.