Lock scroll wheel testing behind "debug" arg
I do some testing with the wheelEvent (as it does not do anything in the program). But should not have this included in the actual releases, so I am locking it behind a command line argument "debug".

In testing if you want to use this simply run the program via
"py AlternativeDesktop.py debug" instead of just "py AlternativeDesktop.py".