Logging for detection of infinite paint loops
Changed print statements which printed every time an icon was painted (for debugging) to a logger which only attaches to the first icon in self.desktop_icons (if it exists). It will log every 100 repaints. Thus if it appears multiple times in a short period of time it is proof of an infinite repainting bug which has been found twice and fixed twice.