Hovering text now shows tooltip with full icon name
Now when hovering over the text area of an icon it will display a tooltip with the full icon name. (useful for small icons if the full icon name does not fit within the lines allocated to it).

Does not appear when hovering over icon itself, just the text area of an icon. And only appears when hovering either directly over text, or in the area in line with a text (not the full text box for icons with only 1 line).