DesktopIcon images now maintain aspect ratio.
Icons will no longer be force scaled into ICON_SIZE x ICON_SIZE square images. It now maintains the aspect ratio of the source.

It also still maintains its entire object as a clickbox, so no changes affecting what constitutes as a click on the icon or not. i.e. you can double click/right click on the now blank sides of a tall and skinny desktop icon and it will still trigger as if it was a square icon.

Gif images are still squares :(