Fixed problem with console not receiving errors.
When setting up to send stderr to the log file, it removed the full error outputs from the base console, requiring loading up the log file to see the actual full error.

Added a TeeStream class which like the linux command splits the stderr to both the console and the log file.