Working on logging problems before release.
Planned to release V0.3.000 last night but ran into massive problems with the logger on the installed version. Mostly surrounding the way that a noconsole has to redirect the stderr. Kept running into errors that would either invalidate the logger setting it to none or just refuse to log anything, or even turn every logging.info/error etc. itself into an error message :).

I have somewhat fixed it, as is right now running it normally through dev environment/installer both seem to work. The problem is this is redirecting all logging and errors to the log file. Thus the console for dev work is pretty much empty. I think I will add a workaround for using the old logging setup for development.