Added dev and devbug cli arguments for logging
Running the program you now have the option to add:
python AlternativeDesktop.py dev       or
python AlternativeDesktop.py devbug       (for debug on scroll wheel up/down).
for development.

This uses the non-redirecting stderr so that everything still appears in the console. Note this mode also removes the errors from the log file (but still shows them in console).

Running with no args or "debug" will test the same version of logging as it should have when installed. All logging should go directly to the log file, including errors, and nothing will show in console.