Moved installer logging setup to logs.py
Moved logging setup to logs.py

Logs calls are as follows:
setup_dev_logging() = Developer mode which does not redirect stderr to logging but to terminal. (Called through arg: "dev" or "devbug"

Logging call for installed version/testing installed version: setup_logging() (called with no args or "debug" arg).

Also changed installed logging version to DEBUG, as it was incorrectly assumed that INFO was lower than DEBUG, which it is not...