Fix gif icons not allowing swapping folders.
Attempting to swap a gif folder would raise a PermissionError as the file would be considered "in use" because of the playback of the gif.

Icons movies now get set to None before attempting to swap their folders, which should allow them to be swapped, and then reloaded in the correct locations, while avoiding such an error.