Stop icons from swapping if swap_folders fails
When swapping icons it now attempts to swap the folders first, if it fails to swap the folders due to an error, do not swap the icons at all, and rollsback the _temp file to normal.

Should effectively not do anything if it fails to swap the folder due to some error. One way to test this is to remove the lines under the comment:
# Stop the movie or it can receive permission errors as a file in use.
Then swap a gif file or swap with a gif file. It should not swap the icons, visually, references, or folders.

Also fixed a presupposition that the swap would work thus assigning the icons row and col to the new_row, and new_col. This worked when attempting the swap regardless if it causes problems with folders/etc. However, it is not needed to change its row/col at the start and will be changed later regardless. Thus the presupposition at the start of the swap (in DesktopIcon's mouseReleaseEvent) instead just sets up old_row, old_col, new_row, and new_col. Which then swap_icons is responsible for all changes instead of swapping in multiple parts.