Media player now pauses before swap_icons.
Due to a time.sleep in the retry of the swap_folders (called by swap_icons) if there is a problem swapping folders, a few frame drops can occur which lags the player for a few frames to a second as it recovers. It is better to avoid all instances that these frame drops occur as logging them is not currently possible. Further, for older machines, prolonged time of dropping frames could likely lead to a cascading failure killing the program (never encountered just a hypothesis).