Scaling video to fit viewport by width
First attempts at allowing more customizing to how the video appears in the program. Starting off with a simple fit_width. Instead of the previous implementation which would fit the whole video in the viewport even if it put blank bars on the sides/top. The ideal is being fully customizable, preferably with sliders maybe even a customizable window or at least sliders etc. to allow a user to place the video exactly how they like it.

For now is hard coded to use "fit_width" as video_scaling_mode setting is not set, and cannot be changed in the program.