fit_width now centers vertically on default
Ideally will have sliders to allow more customization how it appears, but for now it will just default to showing centered vertically.

Some videos which were wide, especially wider than 21:9 ratio  (2292x960 video) would have a horizontal gap at the top of the viewport. In 1920x1080 (16:9) it would still fit but have the gap at the top. In 344x1440 (21:9) it would have the gap at the top and not even close to fitting (~30-50% of the video missing).
Afterward, 1920x1080 video has some room on top and bottom but is centered in the viewport. 3440x1440 is missing some on top and bottom, but is centered and as long as the center part is the focal point it is more intuitive. (As stated something like a slider to adjust the height/add extra width etc. to allow more customizing to fit everything or just the import parts of a video).