Rough, zooming in/out on video bg.
Not ideal, just a proof of concept. Will probably have to do something like a point that it follows and zooms based on center of the screen, not like this where it scales into the left side.

Changed scroll wheel up/down to only zooming for now (both in and out of debug mode). Need to get this stuff settled and then will add actual settings for it.

Spent like 4-5 hours trying to improve the performance of the QGraphicsVideoItem, by either replacing it with a QVideoWidget (incredibly efficient compared to the sloppy mess of a QGraphicsVideoItem). But it is seemingly impossible to add the QVideoWidget to be behind the QGraphicsView, ProxyWidgets don't work, QGraphicsView with transparent backgrounds still cover over the QVideoWidget etc. It seems pretty much impossible, which is super annoying because QVideoWidget uses ~30-40% the ram, and 0-2% CPU vs QGraphicsVideoItem which uses 5-10% CPU (4k 60fps 30kbr video). It only uses this much cpu in the foreground at least, but it is still annoying that it seems impossible to improve. Even tried to work with cv2 or ffmpeg and it was either the same performance or caused memory leaks in literally the gigabytes per second....