Fix tall videos not correctly scaling to fit width.
Had a problem getting tall videos to scale to fit width. Would only actually scale if calling self.scale_to_fit_width() afterwards in something like wheelEvent. Found out the reason was that it was not correctly getting the aspect ratio due to trying too quickly to load it. a retry with timer allows it to resize it correctly once it is properly loaded. Also removed the self.vertical_bg to try some stuff with the video zoom/moving, like running this purely as an init and not a run after changing zoom/move function.