Making video_item into a class VideoBackgroundManager
Separating out the background video into a class to make it easier to work with. For now only zooming in and out based on a center_x, center_y point works. (bound to scroll up/down for now).

Also removed reference to fit_width or scale_to_fit_width as it is more important to get the actual moving and zooming it correctly first then try and do a default transformation on a new background video. For the time being it will just show the video as it is normally appears (scaling by aspect ratio until height or width takes up the whole screen, then stopping).