Now draws center point in debug or debug mode.
Added a visual cue to the center point which displays in debug or devbug (python alternativedesktop.py devbug) mode. This should theoretically always stay in the very center of the screen as functions to move it around come into existence, but being able to visually see it is nice for debugging as it should immediately tell you something is wrong if it is not in the center.