Add timer to ensure proper aspect_ratio return.
Added a timer which recalls init_center_point until a valid response is returned from get_video_aspect_ratio (to ensure that a valid aspect ratio, and video_height and video_width are returned).

Before it was just using an assumed 500ms that it would work by then, but now it attempts it every 50ms until it returns with a valid aspect ratio or at a maximum of 50 retries for 2.5 seconds (50 retries x 50ms in between = 2500 = 2.5 seconds).

All testing generally takes ~14 retries before it returns (~0.7seconds) For now it is not the end of the world as aspect ratio and width/height is not incredibly important. But also of note is that you cannot instantly call init_center_point it also needs a delay of some time or else it will just return center_x, center_y as 0,0 (calls it before it is properly resized) so this is the reason the timer for get_video_aspect_ratio is also in init_center_point and not another function.