Moving the video_item
Can now call move_video(x_offset, y_offset). These are static values which set the offset. Zooming is properly independent, such that zooming then moving still puts the focus in the same area as moving then zooming. Zooming will be changed to static value like move_video, but for debug its nice to be able to zoom in additively.

I was wrong when I said that the center_point would be good just by seeing if the center_point was off of center, as the center_point literally doesn't move, unless you resize the window it should always be in the very center of the viewport. It is helpful for testing exactly what you should be zooming in on however so I am keeping it in debug (and for resizing support eventually).

Not sure exactly how to approach resizing in general as it is an added complexity programs like this usually don't have. Will try just updating the center point and see how things work and if anything needs to change.