Resizing now updates center_point location
Resizing keeps the center_x,center_y in the center of the screen so zooms are based on that location.

Only problem is that the viewport should likely resize to keep the center points target essentially the same. So you would then lose out on area to the sides/above/below, but the center focus object of the video would appear the same.