Update move_video to use floats/percentages
Center point is now decided upon by a percentage value, thus resizing keeps the center point in the same location.

move_video() now uses float values for x_offset/y_offset with -1, -1 cooresponding to the bottom left corner in the focus point, 0, 0 being very center in focus, and 1, 1 being the top right of the video in focus. These can go much further than 1/-1, but when implementing the settings should probably keep it reasonable like either 1 or at max like 2 or 2.5.

The awkwardness of this commit: Since the settings themselves are not defined/set, when resizing after moving for now just do another scroll up (this move_video() call is essentially what would be ran but with the settings). See line 137 for example.

Essentially launch program, scroll down to zoom, scroll up to go to (0.5, 1) (top mid-right). Resize the program, the point should not be in the same location, scroll up to get the same effect as if it had a setting bringing it back to the same top mid-right focus point. It's wonky because there are no settings for it, this is just a proof of concept of how it will work with a roundabout way of proving it.