Added video_x_offset, video_y_offset, video_zoom settings
Added settings:
video_x_offset: x_offset scaler (-1 to 1) covers left edge of the video to right edge.
video_y_offset: y_offset scaler (-1 to 1) covers bottom edge of the video to top edge of the video.
video_zoom: Not used at all yet but will be a static scaling amount of zoom.

Settings will likely go below -1 and above 1, but -1 to 1 is the full scale of the focus being within the dimensions of the video.

Upon resizing it performs move_video to keep the video's focus as the center of the screen, thus making the screen smaller cuts off the edges/top/bottom. Making it larget adds additional height/width. Zoom remains the same.