Added setting: Video horizontal adjustment.
Can go through -1.50 and 1.50 (slider values are ints -150 to 150 then just float divided to get the scaling it expects).

Video position is previewed, and reverted upon cancelling a save.

Known issue, the move_video in resizeEvent() is calling too quickly such that when launching the program with a video_x_offset, it will still zoom based on the direct center of the video and not the center of the viewport. This changes whenever move_video is called after inititial resizing, (like changing in settings or the scroll up for now).