Move_video now called from init_center_point on resize
Fixed the bug mentioned in last commit about how when move_video is called in resizeEvent on program start it still zooms in on the center of the video, not the center_point.

This now gets called after init_center_point thus should correctly zoom in on the center point set immediately after launch. (Still works as expected after changing and saving, or cancelling a save.)