Added Video zoom adjustment in settings menu.
Because it is a bit complex, I have split this into 2 commits. This first part is just making sure that it converts the "video_zoom" setting to a value 0-200, with 0 being zoomed maximum amount out (0.15 zoom factor), 100 being default/normal zoom (1.00 zoom factor), and 200 being max zoomed in (15.0 zoom factor).

Set up video_zoom_to_slider to possibly allow changing the min_zoom or max_zoom to beyond 0.15 and 15.0 limits (unlikely to do as the slider is limited to 201 possible values so say changing the max_zoom to 100, means you are zooming in a lot more per increase in zoom >100, 15.0 is already a very far zoom factor unlikely to be used in almost all scenarios.