Better non-linear scaling system for video zoom.
Using a linear scaling system when zooming out seems fine. However a linear scaling system when zooming in is quite a bit of zoom per each slider tick.
To fix this it would either be to lower the max_zoom, thus you won't be able to zoom as far, or make it non-linear.
Opting for the non-linear for slider > 100 (zoom >1.00) works quite well, a zoom in for a 16:9 video to fit a 21:9 monitor now has the value 111 instead of 103. Therefore each zoom value close to 100 (where I assume 99.9% of zooms will be) allows for much more fine tuning.
It also feels more linearly (even though its not) with the zoom as before zooming into 100 was linear, but then it would zoom in quite a bit following 101+.

If anything I might look at making it scale a little bit less around 100 for a little bit more fine tuning.