Improved non-linear zoom scaling.
It has become more non-linear.
Scaling after 100 is even smaller adjustments, slowly making slightly larger adjustments up to 200 where it reaches max_zoom.

For instance in linear zoom, a 21:9 screen to zoom in to fill the screen with a 16:9 video. A zoom value of only 103 would fill the screen, therefore every zoom value meant a LOT of zooming was being done.
With the non-linear exponential scaling value 1.5 (previous value) it took a zoom value of 111 to fill the screen.
With this new 2.1 exponential scaling value it takes a value of 120 to fill the screen. This is good because it means each zoom value is much smaller thus it is much easier to have a zoom that better fits the window without cutting off a large part of the sides or top just to not have the blank bars on the sides/top.