Fix multiple center_dots in debug mode
Stops the center_dot from "bleeding" on drag resizing. (Redrawing again and again creating a line of dots).

Mostly went unnoticed because originally it never had the possibility for multiple, then when trying to get it to work with resizing it was allowed to create multiple, but wasn't removing old ones from the scene.