Primitive resizable settings window.
A somewhat basic QScrollArea and maximum size set. I call it primitive because it just uses a hard coded self.resize(600, 800) for initially trying to set the width/height, vs the old version which would just give you only the version where everything fit nicely (at the expense of it possibly not fitting).