Better settings_menu sizing
Instead of the primitive directly setting width/height of the resize it now attempts to resize it to the normal dimensions it would take up.

If this size is too large, it gets downsized to fit the screen via the setMaximumSize, which will cut the height to screen height -40 (to account for taskbar), and cut the width to fit the monitors width.

The menu is still resizable and has a scroll area, just the default size it appears as better reflects the size it should be.