Hiding video slider if video bg not possible
Basic attempt at hiding a row for one of the video adjustments if the background setting is set to a something that does not even allow the possibility for a video background.
I.e. hide the row if background option = "Image only" or "None"

For now only adds/hides the video horizontal adjustment. (Testing some stuff before changing it for all 3 of them.)