Function for resizing on adding rows.
Since there is now rows that might not appear when the settings menu is loaded if adding the rows in update_video_slider_visibility  can now call resize_window(width, height) which will add on additional room from the base width/height it appeared as.

Was having some trouble with just running the same calculation as at the end of the init, ran into some trouble with it where it kept trying to make it skinnier, and calculations were based off of if the layout was already showing not whether it was just told to show() or hide() in the update_video_sliders_visibility().

Not sure if this will stay, as either a tab system which had been planned or something similar will probably make this a lot more annoying.