Better resizing on adding/removing rows in Settings
Found the problems mentioned in last commit with the calculations. Seemed to be a difference in the raw .width() or .height() after init. Moved it over to .sizeHint().width() and .sizeHint().height() which seems to update correctly.

Now fully sets the size of the window to fit the content that shows when self.resize_window() is called.
(for example: when video adjustments are shown it resizes to fit them, when they are hidden it resizes the window to fit what is shown.)