Settings menu no longer defaults to label color box
When hitting enter upon opening settings it would auto default to the color box. This was originally a problem with hitting enter auto saving as the save button would be set to default, but was changed to prevent that. Simply did not apply it to the color box when added.