Fix bad reference preventing keybind change
In putting the QDialog options inside the content_widget it essentially made self.parent() reference the content_widget as it's parent instead of the SettingsDialog as a whole, thus content_widget did not have a set_changed() function and would not save/display the keybind.

KeybindButton is now passed a reference to the dialog so that it can call self.dialog.set_changed() directly.