Fixed background paths folder buttons.
These can now be selected while tabbing through, and respond to "enter" key. Also disabled focusing of the ClearableLineEdit "x".

Next focus will be a 2nd or 3rd tab to separate a few settings. Or some clarification on what is selected when tabbing through (some thing like the sliders do not give an indication when selected, but just look normal).