Theme and selection stylesheet tweaks.
Desktop.py now stores an up to date version of the theme_colors. For instance changing to None theme now sets all colors to None. Previously it would have just retained the old theme.

Settings_menu.py:
The focus stylesheet from last commit now applies to whole settings_menu (applies to the SliderWithInput classes as well).

When theme_selector is None it uses the default stylesheet, and when it is NOT None it uses the focus stylesheet. This is set up such that on swapping it will also change the stylesheet thus not needing a relaunch of settings to apply.
^ The whole reason this is implemented is that by just changing the stylesheet it was changing the way the "None" theme worked as well (deleting the color of the sliders leaving just a border color).