Fixed focus for color box
The "Icon Name color" color selection box now has a focus style, of 3px of the theme's color applied. This follows the same stylesheet application enable_stylesheet() so it also only applies when theme is "Dark" or "Light" and disabled when theme is set to "None".

Removed update_color_box as it's use is no longer needed and the stylesheet is updated via enable_stylesheet() function.