Fix Icon name color not updating after changing.
Fixed the box to now change color after updating.

Also in doing so found a critical bug (With no traceback or errors, my favorite type :) ).

How to trigger: change themes twice (from dark to light save, then back to dark and crashes before it changes theme). Precisely it fails at the apply_stylesheet lines in apply_theme. The older (current live version of the project) settings menu does not have this problem. So it is likely just a tracking down of the difference to find out the problem.