Fixed silent crashing on changing theme twice.
Described briefly in last commit, but will restate problem here:
Program would silent crash on swapping Theme twice.
Could reproduce it by launching, changing theme to dark or light, saving, then changing it again (would crash before changing theme).

I thought it was the apply_stylesheet as when trying this line in desktop.py it would crash.
Tracing down every change from the last stable release, I found that the KeybindButton being passed a reference to it's dialog parent (a quick hack to allow it to call set_changed() with the reformatting into a scrollable area) was causing the problem.

Direct referencing is in hindsight an incredibly dumb idea but something that worked in the moment just for checking viability, and would ironically work as long as apply_stylesheet was never called again after initialization.

Not 100% sure on the exact error, but I assume it is a cascading error from apply_spreadsheet trying to update KeybindButton which then tries to update it's dialog reference, which then tries to update KeybindButton and so on...

Silly quick fix caused a good 30+ minute bug chase, and would have been solved by a literally even quicker fix.... (1 line)