Resize_window() keeps menu on screen.
Whenever resize_window() window is called, it now checks if the height of the screen would overflow below the bottom of the screen and adjusts the Y location of the menu such that the entire menu stays on screen.

This should always work without having problems like it moving it too far up, as the setMaximumSize should be guaranteed to be the at max the height of the screen -40, but added a check which will not move the window if it would attempt to move it to a negative Y height (above top of screen).