Custom bg color previews
Now when fully selected a color from the color picker it previews the fill color.

Also previews the fill color based on custom_bg_fill setting. For example turning it off previews the default background color.
In other words when you are ready to click save, what you see is what it should be.