Custom bg color picker now previews color
When in the color picker for the custom bg color, changing the color now gives a live preview of what it would look like.

For instance you can now use the "pick screen color" to live preview any color on your screen. Perfect for matching the sides or top/bottom of a video background for when the video doesn't perfectly fill the screen.