Custom bg color selector now reverts if no color selected
Now reverts to the color when the menu was opened if the user either X closes or hits cancel to close out of the color picker.

Previously it was updating the self.bg_color whenever it changed, AND also wouldn't re-call render_bg() thus would keep the same previewed color as the last previewed color in the selection dialog.

Also removed redraw request for this color change as it should not require a redraw, all it requires is a render_bg which it calls itself and on close/save by default. And redraw request is more taxing.