Image background maintains focus on resizing
The focus point (center of the image + "Image horizontal adjustment" and + "Image vertical adjustment") now always stays in the center of the viewport, even among resizing.

This does mean that the center point does move upon resizing, but maintains that the center of the viewport will always be the focus point (Ideal situation).

Also noticed a problem with how the video object is working, during testing it was only tested with full screen, then setting it to normal state, however it was noticed that when resizing from normal state it does resize the video object (instead of just cutting off/showing more). Will look to addressing this at some point.