Step towards static video background
Moving to try and get the video background to work like image background, especially on resizing. For instance the old videobackgroundmanager has some lasting problems due to how the video_item's size is adjusted upon resizing the window. This causes some problems with the video_item attempting to fit to match the screen, thus changing the size of itself and causing problems like making the window skinnier also makes more vertical height of the video fit within the same view, instead of just cutting off the areas that should no longer render it instead resizes the height of the video as well.

For now it is kind of broken as in the center point of the video object does not correspond correctly and there are quite a few problems that need solving, but the video object itself is a static size of the same height/width of the actual video input.