VideoBackgroundManager now works like ImageBackgroundManager
The size of the video_item is now static at the same resolution as the video file.

Resizing has been updated to now match the system that ImageBackgroundManager uses.

Since video_item is now static size it no longer resizes upon being too skinny/too short (this was a limitation of just allowing the QGraphicsVideoItem control how it originally appears inside the viewport).

known issue: debug mode center point, currently paints it at the center of the video_item NOT the center of the viewport (Zooming still works around the center of the viewport however so this is just a visual bug which will be fixed).